data/requirements.txt
Flask==1.1.1
Pillow==7.1.1
numpy==1.17.0
